Add automation for setting up server
A new `sh` script has been added that can download the necessary server files and the Skyblock Mini datapack and gets
the server set up for the user. Also generated is a script that starts the server with Aikar flags. Use of the script
has been documented in the README.

The manual server setup has also been updated in the README in order to simplify the process. E.g., only a single server
start is needed instead of three.