Split shrieker fill command into separate command per y-level
This avoids the issue of the entire command failing just because one coordinate is too low or too high.

We also constrain the fill volume to only 6 blocks from the player instead of 15. This is running every tick, after all,
and a radius of 15 corresponds to almost 30 000 blocks being filled each tick!