Add a crafting recipe for summoning sculk shriekers
The recipe takes a sculk shrieker surrounded by echo shards and outputs another sculk shriekers with the `can_summon`
block state set to `true`. This makes it able to summon the warden. There is no other way to obtain such a shrieker as
an item.

In order for this implementation to function, the `nbt-crafting` mod is required. It has been added to the setup
instructions and automated setup scripts. If for some reason the mod is not present, the crafting recipe will cause an
error to be logged but will not otherwise interfere with the game (so there will only be natural sculk shriekers that
can summon).

Sculk shrieker blocks are no longer automatically converted to summoning sculk shriekers as that functionality is now
accessible through a crafting recipe.