Remove the need for a starting chest
Instead of including a lava bucket and an ice block in the starting chest, we now embed a lava source block and a water
source block in the island itself. This suffices to build a cobblestone generator, but the lava and water are not
movable until iron is obtained for a bucket, nor can the water be duplicated until a glass bottle or dripstone is
obtained.

The mycelium was also removed from the starting chest. Instead, nine blocks of mycelium now form the floor of the allay
cage in the end. Finally, a new chest was added in the nether which is meant to include some early-to-mid-game
items. For now, it only contains cocoa beans which serve the purpose of acting as a dye, or if desired for making
cookies. With nothing remaining in the starting chest, it was removed from the spawn island.

The README and the spawn island screenshot have been updated to match.